FU-54 refactor: jwt 패키지 분리
